refactor: extract platformName
