+++ src/schemas/platformName.ts
@@ -1,0 +1,16 @@
+import { z } from 'zod'
+
+export const platformName = z.enum([
+  'cake',
+  'yourator',
+  '104',
+  '518',
+  '1111',
+  'chickpt',
+  'meetJobs',
+  'taiwanJobs',
+])
+
+export type PlatformName = z.infer<typeof platformName>
+
+export const platformNames = platformName.options
